Fix selections in input, replace DOuble with BigDecimal

+++ app/src/main/java/com/minimalisticapps/priceconverter/data/remote/dto/BitPayExchangeRate.kt
@@ -4,24 +4,31 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 @Entity(tableName = "Coins")
 data class BitPayExchangeRate(
     @ColumnInfo(name = "code")
     var code: String,
+
     @ColumnInfo(name = "rate")
     @SerializedName("value")
-    var rate: Double?,
+    var rate: BigDecimal?,
+
     @PrimaryKey(autoGenerate = false)
     @ColumnInfo(name = "name")
     val name: String,
+
     @ColumnInfo(name = "type")
     val type: String,
+
     @ColumnInfo(name = "unit")
     @SerializedName("unit")
     val unit: String,
+
     @ColumnInfo(name = "one_shit_coin_value")
-    var oneShitCoinValue: Double?,
+    var oneShitCoinValue: BigDecimal?,
+
     @ColumnInfo(name = "one_shit_coin_value_string")
     var oneShitCoinValueString: String
 )
